requirements.txt
django~=2.1.7
pillow
djangorestframework
pygments
